refactor!: migrate auth feature to its own feature folder
