+++ server/src/features/auth/controllers/handle-logout.ts
