+++ server/src/features/auth/index.ts
@@ -1,0 +1,6 @@
+export * from "./controllers";
+export * from "./schemas";
+export * from "./services";
+export * from "./utils";
+export * from "./auth.middleware";
+export * from "./auth.routes";
