+++ server/src/features/auth/schemas/register.schema.ts
