+++ server/src/features/auth/services/index.ts
@@ -1,0 +1,2 @@
+export * from "./repositories";
+export * from "./user-data.service";
