+++ server/src/features/auth/services/repositories/index.ts
@@ -1,0 +1,2 @@
+export * from "./user-role.repository";
+export * from "./user.repository";
