+++ server/src/features/auth/types/auth.d.ts
@@ -1,0 +1,11 @@
+import { UserDataService } from "@/services";
+import { TokenPayload } from "../schemas";
+
+declare global {
+  namespace Express {
+    interface Request {
+      userDataService: UserDataService;
+      authPayload?: TokenPayload;
+    }
+  }
+}
