+++ server/src/routers/index.ts
@@ -1,5 +1,4 @@
-import authRouter from "./auth.router";
 import projectsRouter from "./projects.router";
 import projectTagsRouter from "./project-tags.router";
 
-export { authRouter, projectsRouter, projectTagsRouter };
+export { projectsRouter, projectTagsRouter };
