+++ server/src/services/data/index.ts
@@ -1,2 +1,1 @@
 export * from "./project-data.service";
-export * from "./user-data.service";
