+++ server/src/services/repositories/index.ts
@@ -1,4 +1,3 @@
 export * from "./abstract.repository";
 export * from "./project.repository";
 export * from "./project-framework.repository";
-export * from "./user.repository";
