+++ server/src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from "./hash/hash.util";
 export * from "./jobs";
 export * from "./logger";
-export * from "./auth";
