+++ server/src/viewmodels/index.ts
@@ -1,3 +1,2 @@
 export * from "./cache";
 export * from "./schema.types";
-export * from "./schemas";
